add working position slider, and on screen information update
The bottom slider now moves to the position in the book where it lands.
The numbers above it move based on the word count, and an ETA is
calculated based on (total words - words read)/Words per minute = minutes
left. The Both sliders update the ETA value when they move so if one
increases the WPM the ETA on  the bottom slider goes down. The word
counter on the bottom also has a percentage to 2 decimal places. The WPM
is loaded onCreate and saved on stop.